refactor(builtins): move OpenAI model client out of inference
tacklr.InferenceStrategy stays the harness interface. The OpenAI
Responses client is a host battery, so it lives in builtins. Package
inference is removed. streaming still owns Message and StreamEvent.
