+++ builtins/openai_capabilities.go
@@ -1,4 +1,4 @@
-package inference
+package builtins
 
 import (
 	"strings"
