+++ builtins/openai_classify_test.go
@@ -1,4 +1,4 @@
-package inference
+package builtins
 
 import (
 	"errors"
