+++ builtins/openai_sse_test.go
@@ -1,4 +1,4 @@
-package inference
+package builtins
 
 import (
 	"context"
