+++ doc.go
@@ -5,7 +5,8 @@
 //   - streaming owns messages, events, tool calls, and todos.
 //   - brain owns knowledge retrieval and graph capabilities.
 //   - vfs owns virtual filesystem mounts, sessions, and provider interfaces.
-//   - builtins owns optional tool constructors (email, Exa) and VFS backend factories.
+//   - builtins owns optional tool constructors (email, Exa), VFS backend
+//     factories, and the OpenAI-compatible model client.
 //   - mcp owns MCP connection configuration.
 //
 // Process-wide registrations (built-in interrupts, common VFS codecs, the
